refactor: extract DeleteTask from ApplicationController
